% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{examplePath}
\alias{examplePath}
\title{Returns path to examples folder}
\usage{
examplePath()
}
\value{
Returns path to examples folder
}
\description{
Returns path to examples folder
}
\examples{
examplePath()
}
