% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WordR.R
\docType{package}
\name{WordR}
\alias{WordR}
\alias{WordR-package}
\title{WordR: package for rendering MS Word documents with R inline code and inserting tables and plots.}
\description{
The WordR package provides three main functions
\code{\link[WordR]{addFlexTables}}, \code{\link[WordR]{addPlots}} and \code{\link[WordR]{renderInlineCode}},
which together should provide enough functionality to create a nice, customised MS Word report from a given template.
}
\details{
More information about the package can be found in \href{https://CRAN.R-project.org/package=WordR/vignettes/WordR.html}{vignette} or run:\cr

\code{vignette("WordR", package = "WordR")}
}
