% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderInlineCode.R
\name{renderInlineCode}
\alias{renderInlineCode}
\title{Read Word document with R code blocks, evaluate them and writes the result into another Word document.}
\usage{
renderInlineCode(docxIn, docxOut, defaultStyle = NA, debug = F)
}
\arguments{
\item{docxIn}{String of length one; path to Word file with R code blocks.}

\item{docxOut}{String of length one; path for output Word file.}

\item{defaultStyle}{String of length one or NA; Default style for inserted text}

\item{debug}{Boolean of length one; If True browser() is called at the beginning of the function}
}
\value{
Path to the rendered Word file if the operation was successfull.
}
\description{
Read Word document with R code blocks, evaluate them and writes the result into another Word document.
}
\examples{
\donttest{
renderInlineCode(
  paste(find.package('WordR'),'examples/templates/template1.docx',sep = '/'),
  paste(find.package('WordR'),'examples/results/result1.docx',sep = '/'))
  }
}
